refactor: Split service_modules.go by concern to reduce hotspot complexity
Decomposes the 679-LOC service_modules.go (file-level aggregate
cyclomatic complexity ~136) into focused single-concern files within
the same package, leaving the public API and behaviour unchanged:

- handler_tree.go: handlerTree type, parsing, and naming-conflict validation
- starlark_convert.go: Starlark<->Go value conversion helpers
- compensation.go: saga step-result tracking and compensation params
- authorization.go: RBAC handler authorization
- context.go: thread-local StarlarkContext storage
- service_modules.go: retains error definitions and core module builders

No function previously exceeded the cyclop per-function threshold of 15;
the hotspot was driven by aggregate complexity and 679 LOC in one file.
This is a pure code-move refactor: build and tests are unchanged.
